+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v10.8.0 (2021-08-04)
+
+* Bumps Instabug native SDKs to v10.8
+* Adds string keys for the discard attachment prompt dialog.
+* Fixes Autolinking on iOS.
+
 ## v10.4.0 (2021-05-10)
 
 * Migrates iOS to use XCFramework